# GoRestAPI/swagger.json
{
  "consumes": [
    "application/json",
    "application/xml"
  ],
  "produces": [
    "application/json",
    "application/xml"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "the purpose of this application is to provide an application\nthat is using plain go to serve a rest api",
    "title": "Snippets",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "contact": {
      "name": "Roither | Seiberl",
      "email": "andi.roither@protonmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/",
  "paths": {
    "/login": {
      "post": {
        "description": "Returns a user for provided email and jwt",
        "produces": [
          "application/json"
        ],
        "operationId": "login",
        "parameters": [
          {
            "type": "string",
            "description": "user id for user selection",
            "in": "body",
            "required": true
          },
          {
            "type": "string",
            "description": "web token from user authentication",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/User"
          },
          "401": {
            "description": "authorization failed"
          },
          "404": {
            "description": "no user with that mail found"
          }
        }
      }
    },
    "/user/{userID}/snippets/": {
      "get": {
        "description": "Returns all snippets for a user",
        "produces": [
          "application/json"
        ],
        "operationId": "getSnippets",
        "parameters": [
          {
            "type": "string",
            "description": "user id for snippet selection",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "authorization token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "HTTP status code 200 and Snippets",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/responses/Snippet"
              }
            }
          },
          "401": {
            "description": "authorization failed"
          },
          "404": {
            "description": "no user(user) with that id found"
          }
        }
      },
      "post": {
        "description": "Create a snippet with a given id",
        "produces": [
          "application/json"
        ],
        "operationId": "createSnippet",
        "parameters": [
          {
            "type": "string",
            "description": "user id for snippet selection",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "authorization token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "snippet id",
            "in": "body"
          },
          {
            "type": "string",
            "description": "snippet name",
            "name": "name",
            "in": "body"
          },
          {
            "type": "string",
            "description": "snippet language",
            "in": "body"
          },
          {
            "type": "string",
            "description": "snippet description",
            "in": "body"
          },
          {
            "type": "string",
            "description": "snippet code text",
            "in": "body"
          }
        ],
        "responses": {
          "200": {
            "description": "an updated snippet to be returned",
            "$ref": "#/responses/Snippet"
          },
          "401": {
            "description": "authorization failed"
          },
          "404": {
            "description": "no snippet with that id found"
          }
        }
      }
    },
    "/user/{userID}/snippets/{snippetID}": {
      "get": {
        "description": "Returns a snippet for a given user and snippet id",
        "produces": [
          "application/json"
        ],
        "operationId": "getSnippet",
        "parameters": [
          {
            "type": "string",
            "description": "user id for snippet selection",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "snippet which should be retrieved",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "authorization token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "a snippet to be returned",
            "$ref": "#/responses/Snippet"
          },
          "401": {
            "description": "authorization failed"
          },
          "404": {
            "description": "no snippet with that id found"
          }
        }
      },
      "put": {
        "description": "Update a snippet with a given id",
        "produces": [
          "application/json"
        ],
        "operationId": "updateSnippet",
        "parameters": [
          {
            "type": "string",
            "description": "user id for snippet selection",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "snippet id",
            "in": "path"
          },
          {
            "type": "string",
            "description": "authorization token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "snippet name",
            "name": "name",
            "in": "body"
          },
          {
            "type": "string",
            "description": "snippet language",
            "in": "body"
          },
          {
            "type": "string",
            "description": "snippet description",
            "in": "body"
          },
          {
            "type": "string",
            "description": "snippet code text",
            "in": "body"
          }
        ],
        "responses": {
          "200": {
            "description": "an updated snippet to be returned",
            "$ref": "#/responses/Snippet"
          },
          "401": {
            "description": "authorization failed"
          },
          "404": {
            "description": "no snippet with that id found"
          }
        }
      },
      "delete": {
        "description": "Deletes a snippet for a given id",
        "produces": [
          "application/json"
        ],
        "operationId": "deleteSnippet",
        "parameters": [
          {
            "type": "string",
            "description": "user id for snippet selection",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "snippet which should be deleted",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "authorization token",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "result success"
          },
          "401": {
            "description": "authorization failed"
          },
          "404": {
            "description": "no snippet with that id found"
          }
        }
      }
    }
  },
  "responses": {
    "Snippet": {
      "description": "Snippet from DB\nHTTP status code 200 and Snippet",
      "headers": {
        "category": {
          "type": "string"
        },
        "code": {
          "type": "string"
        },
        "language": {
          "type": "string"
        },
        "owner": {
          "type": "string"
        },
        "snippet_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        }
      }
    },
    "User": {
      "description": "User from DB\nHTTP status code 200 and User",
      "headers": {
        "mail": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "user_id": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    }
  }
}